Day 14, addendum
This is a better approach. Rather than make one unit of fuel at
a time, set the target to 'n' units of fuel and work out the ore
requirement to make that.

If the ore requirement is within the target amount, great.
Otherwise, try to make less fuel. Repeat with a binary search.